+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Add "Export all" and "Import all" general actions
 * Add `import_duplicate_moniker_creature` setting, which defaults to false
+* Added `Egg` tab, including hatch, pause & resume actions